Add files via upload
This project is written in C and using OpenGL (Open Graphics Library). It is 2D representation of how electrons are arranged in energy levels, or shells, around the nucleus of an atom and can visualize the animation of how the electrons are revolving around the nucleus in their respective orbits.